# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JAC Rental Booking Service API",
        "description": "Standalone microservice handling all rental property booking operations for the JAC platform. Authentication uses Bearer tokens issued by the JAC monolith (Laravel Passport).",
        "contact": {
            "email": "dev@jac.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "Current Environment"
        }
    ],
    "paths": {
...
        "/api/v1/cities": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List cities",
                "description": "Returns active cities. Filter by country_id.",
                "operationId": "cd82a2353546bd942448bcf9f1810ff3",
                "parameters": [
                    {
                        "name": "country_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/City"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/neighborhoods": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List neighborhoods",
                "description": "Returns active neighborhoods. Filter by city_id.",
                "operationId": "eb8f85eadb733c6ff153af0c3b7380de",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of neighborhoods",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Neighborhood"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Browse active rental properties",
                "description": "Returns a paginated list of active properties. Supports filtering by city, neighborhood, type, and star rating. No authentication required.",
                "operationId": "6b260bd627ea9529982b5e2d22043f25",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "Filter by city",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "neighborhood_id",
                        "in": "query",
                        "description": "Filter by neighborhood",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "Filter by property type",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "star_rating",
                        "in": "query",
                        "description": "Minimum star rating",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 5,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_size": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "offset": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "properties": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PropertySummary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get property details",
                "description": "Returns full details for a single property including units, reviews, amenities, and settings.",
                "operationId": "baec64dd63122f15d66b967ca7d21b08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropertyDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "City": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "country_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sanaa"
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Neighborhood": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Haddah"
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PropertyDetail": {
                "description": "Full property details including units, org, and settings",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sea View Villa"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avg_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "star_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "logo": {
                        "type": "string",
                        "example": "properties/villa1.jpg",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "org": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Royal Stays"
                            },
                            "settings": {
                                "properties": {
                                    "cancellation_policy": {
                                        "type": "string",
                                        "example": "moderate"
                                    },
                                    "check_in_time": {
                                        "type": "string",
                                        "example": "14:00:00"
                                    },
                                    "check_out_time": {
                                        "type": "string",
                                        "example": "11:00:00"
                                    },
                                    "allow_instant_booking": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "units": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UnitSummary"
                        }
                    },
                    "rules": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PropertySummary": {
                "description": "Compact property representation returned in list views",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sea View Villa"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avg_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "star_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "logo": {
                        "type": "string",
                        "example": "properties/villa1.jpg",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "UnitSummary": {
                "description": "Compact unit representation in property listings",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Chalet 101"
                    },
                    "max_guests": {
                        "type": "integer",
                        "example": 6
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport Bearer token issued by the JAC monolith. Obtain via the monolith /api/v1/auth/login endpoint.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}